vndb Scraper
Date:?
My first try at web-scraping with selenium. Not very practical since the site has an api for queries, but it was a nice segway to my next scraping project as I started to understand how elements on a page work, and finding what scripts/srcs return different things for the site.