Aperiodic Monotile in python Turtles
this is the new Einstein tile, drawn with an L-System and turtles. Being made with an l-system *hypothetically* would easily allow recursive iterations (Fractals) to be drawn from the shape. I have already made various well known fractals, and many many unnamed shapes with this l-system as a base. More on the system and fractals on my page